Throw errors if the SQL query fails
