+++ src/executeSQL.ts
@@ -1,6 +1,4 @@
-import { open } from "@raycast/api";
-import { showFailureToast } from "./showFailureToast";
-import { baseExecuteSQL, getSystemPreferencesInfo, isPermissionError } from "./sql-utils";
+import { baseExecuteSQL } from "./sql-utils";
 
 /**
  * Executes a SQL query on a local SQLite database and returns the query result in JSON format.
@@ -11,39 +9,26 @@ import { baseExecuteSQL, getSystemPreferencesInfo, isPermissionError } from "./s
  *
  * @example
  * ```typescript
- * import { executeSQL } from "./executeSQL";
+ * import { closeMainWindow, Clipboard } from "@raycast/api";
+ * import { executeSQL } from "@raycast/utils";
  *
- * type User = { id: number, name: string };
+ * type Message = { body: string; code: string };
  *
- * const dbPath = "/path/to/database.sqlite";
- * const query = "SELECT id, name FROM users WHERE age > 18";
+ * const DB_PATH = "/path/to/chat.db";
  *
- * try {
- *   const results = await executeSQL<User>(dbPath, query);
- *   console.log(results);
- * } catch (error) {
- *   console.error("Error executing SQL:", error);
+ * export default async function Command() {
+ *   const query = `SELECT body, code FROM ...`
+ *
+ *   const messages = await executeSQL<Message>(DB_PATH, query);
+ *
+ *   if (messages.length > 0) {
+ *     const latestCode = messages[0].code;
+ *     await Clipboard.paste(latestCode);
+ *     await closeMainWindow();
+ *   }
  * }
  * ```
  */
-export async function executeSQL<T = unknown>(databasePath: string, query: string) {
-  try {
-    return await baseExecuteSQL<T>(databasePath, query);
-  } catch (error) {
-    if (isPermissionError(error)) {
-      const { preferencesName, action } = getSystemPreferencesInfo();
-      await showFailureToast(error, {
-        title: "Raycast needs full disk access",
-        message: `You can revert this access in "${preferencesName}" whenever you want.`,
-        primaryAction: {
-          title: action.title,
-          onAction() {
-            open(action.target);
-          },
-        },
-      });
-    } else {
-      await showFailureToast(error, { title: "Cannot execute SQL" });
-    }
-  }
+export function executeSQL<T = unknown>(databasePath: string, query: string) {
+  return baseExecuteSQL<T>(databasePath, query);
 }
